Start reimplementation of FittableContainer
